+++ changelog.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### fix
 - style issue on toolbar user module.
+- showing unhandable tray icons (ex: nvidia old control panel).
 
 ## [2.1.0]
 ### features